feat: Atualizar pastas de actions
